Conflicts:
	alljoyn_core/src/AllJoynPeerObj.cc
	alljoyn_core/src/AllJoynPeerObj.h
	alljoyn_core/src/KeyExchanger.cc
	alljoyn_core/src/KeyExchanger.h
	alljoyn_core/src/KeyStore.cc
	alljoyn_core/src/PeerState.h
	alljoyn_core/src/Status.xml
	alljoyn_core/unit_test/AuthListenerECDHETest.cc
	alljoyn_core/unit_test/KeyStoreTest.cc
	common/inc/qcc/CertificateECC.h

Additional fixes included in this commit:
  * ASACORE-2003: More instances of ECC key structs being cast to
    byte arrays found and fixed. (KeyExchanger.cc and
    PermissionMgmtObj.cc)
  * ASACORE-2068: Some missed messages discovered and now included
    in the conversation hash
  * ASACORE-2068: Since conversation hash now captures visible
    network traffic only, PSK cannot be included in the hash because
    if they're different between client and server, the running
    digests will permanently diverge and other attempted auth
    mechanisms attempted will be unable to succeed. PSK data is now
    fed separately into the PRF which generates the verifier for PSK
    authentication.
  * AddSignal is now deprecated but appears in a unit_test in the
    feature/security2.0 branch in one instance; this instance is now
    surrounded by warning suppression so it doesn't cause a build
    break. (PermissionMgmtTest.cc)

+++ common/src/CryptoECCMath.cc
@@ -791,7 +791,7 @@ static void big_sub(bigval_t* tgt, bigval_t const* a, bigval_t const* b)
     v = 1; /* increment */
     for (i = 0; i < BIGLEN; ++i) {
         v += a->data[i];
-        v += ~b->data[i]; /* 1's complement */
+        v += (uint64_t)((uint32_t)(~b->data[i])); /* 1's complement */
         tgt->data[i] = (uint32_t)v;
         v >>= 32;
     }
